Work towards defining an API for constructing parsers/automata.
It looks very PEGish right now, though the ultimate emphasis should be on composing pattern recognizing machines together (with unordered choice, in particular) in a way that does not degrade performance at scale.  It should be possible to string together thousands of parsers together with unordered choice and have the dispatch from the first input symbol be a very fast constant-time lookup into a small table.